+++ packages/core/src/element.ts
@@ -1,5 +1,6 @@
 import Action from './action';
 import type { PropertyConstructor, PropertyType } from './decorators/property';
+import { domReady } from './helpers/dom';
 import { camelize, dasherize, parseJSON } from './helpers/string';
 import Property from './property';
 import Store from './store';
@@ -124,14 +125,4 @@ function attributeValueTransformer(_newValue: string | null, _oldValue: string |
     default:
       return { newValue: _newValue, oldValue: _oldValue };
   }
-}
-
-function domReady() {
-  return new Promise<void>((resolve) => {
-    if (document.readyState === 'loading') {
-      document.addEventListener('DOMContentLoaded', () => resolve());
-    } else {
-      resolve();
-    }
-  });
 }