+++ packages/core/src/index.ts
@@ -3,3 +3,4 @@ export { default as registerElement } from './decorators/register_element';
 export { default as target } from './decorators/target';
 export { default as targets } from './decorators/targets';
 export { default as ImpulseElement } from './element';
+export { default as lazyImport } from './lazy_import';