+++ packages/core/src/observers/element_observer.ts
@@ -9,7 +9,7 @@ export default class ElementObserver {
   private started = false;
 
   constructor(
-    private readonly element: Element,
+    private readonly element: Element | Document,
     private delegate: ElementObserverDelegate,
     private readonly observerOptions?: MutationObserverInit
   ) {